Scalatest and loggings fixes to run on 2.13
