+++ blended.updater.config/shared/src/test/scala/blended/updater/config/PrickleSpec.scala
@@ -2,9 +2,10 @@ package blended.updater.config
 
 import java.util.UUID
 
+import blended.util.logging.Logger
 import blended.updater.config.json.PrickleProtocol._
 import org.scalacheck.Arbitrary
-import org.scalatest.prop.PropertyChecks
+import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
 import org.scalatest.{FreeSpec, Matchers}
 import prickle._
 
@@ -12,9 +13,9 @@ import scala.reflect.{ClassTag, classTag}
 import scala.util.Success
 import scala.util.control.NonFatal
 
-class PrickleSpec extends FreeSpec with Matchers with PropertyChecks {
+class PrickleSpec extends FreeSpec with Matchers with ScalaCheckPropertyChecks {
 
-  private[this] val log = org.log4s.getLogger
+  private[this] val log = Logger[PrickleSpec]
 
   "Prickle real world test cases" - {
     "1. deserialize a container info" in logException {
