+++ blended.updater.config/shared/src/test/scala/blended/updater/config/ProfileSpec.scala
@@ -1,9 +1,9 @@
 package blended.updater.config
 
 import org.scalatest.FreeSpec
-import org.scalatest.prop.PropertyChecks
+import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
 
-class ProfileSpec extends FreeSpec with PropertyChecks {
+class ProfileSpec extends FreeSpec with ScalaCheckPropertyChecks {
 
   import TestData._
