+++ blended.util.logging/shared/src/main/scala/blended/util/logging/LogLevel.scala
@@ -1,0 +1,7 @@
+package blended.util.logging
+
+object LogLevel extends Enumeration {
+
+  type LogLevel = Value
+  val Error, Warn, Info, Debug, Trace = Value
+}
