+++ blended.util.logging/shared/src/main/scala/blended/util/logging/LoggerNoOp.scala
@@ -1,0 +1,6 @@
+package blended.util.logging
+
+/**
+ * A Logger class doing nothing.
+ */
+class LoggerNoOp(override val name : String) extends Logger
